requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.59
botocore==1.29.141
Django==4.1.2
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
importlib-metadata==5.0.0
jmespath==1.0.1
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.5
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.16
zipp==3.10.0
